Adjust movementInput vector if its magnitude is too high
Move diagonaly with keybaord input results in having a faster movement than with joysticks.

This is because input values from both axes are set to 1 on keyboard, but around 0.7 for joysticks.

In this particular case, the vector's magnitude is higher with keyboard input. A solution is to divide the movement's vector by its own magnitude, which achieves a similar movement feel for both input types.